move button logic to class
